requirements.txt
echo "python-telegram-bot==13.15
pyTelegramBotAPI==4.12.0
openai==1.30.1
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.7
faiss-cpu==1.8.0" > requirements.txt